+++ src/main/java/com/globallogic/demo/repositories/IPhoneRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.globallogic.demo.model.Phone;
 
 @Repository
-public interface IPhoneRepository extends CrudRepository<Phone, Long>{
+public interface IPhoneRepository extends CrudRepository<Phone, Integer>{
 
 }
